\name{cor.P}
\alias{cor.P}

\title{
Get the Pearson correlation coefficient estimate
}

\description{
Accessor function for Pearson correlation slot in \code{TT.Seq.Gene} or 
\code{TT.Seq.Tx} objects
}

\usage{
cor.P(object)
}

\arguments{
  \item{object}{object of class \code{TT.Seq.Gene} or \code{TT.Seq.Tx}}
}

\value{
Returns Pearson correlation estimates as an vector
}

\author{
Paul K. Korir
}

\examples{
# location of example input files
# to be added...
}

\keyword{correlation}
