\name{cor.S}
\alias{cor.S}

\title{
Get the Spearman correlation coefficient estimate
}

\description{
Accessor function for Spearman correlation slot in \code{TT.Seq.Gene} or 
\code{TT.Seq.Tx} objects
}

\usage{
cor.S(object)
}

\arguments{
  \item{object}{object of class \code{TT.Seq.Gene} or \code{TT.Seq.Tx}}
}

\value{
Returns Spearman correlation estimates as an vector
}

\author{
Paul K. Korir
}

\examples{
# location of example input files
# to be added...
}

\keyword{correlation}
