\name{cors.P}
\alias{cors.P}

\title{
Get the Pearson correlation coefficient estimates
}

\description{
Function that collates all Pearson correlation estimates from a list of objects
of class \code{TT.Seq.Gene} or \code{TT.Seq.Tx}
}

\usage{
cors.P(tt.seq)
}

\arguments{
  \item{tt.seq}{a list with objects of class \code{TT.Seq.Gene} or 
 	\code{TT.Seq.Tx}}
}

\value{
Returns Pearson correlation estimates as an vector
}

\author{
Paul K. Korir
}

\examples{
# location of example input files
# to be added...
}

\keyword{correlation}
