\name{cors.S}
\alias{cors.S}

\title{
Get the Spearman correlation coefficient estimates
}

\description{
Function that collates all Spearman correlation estimates from a list of objects
of class \code{TT.Seq.Gene} or \code{TT.Seq.Tx}
}

\usage{
cors.S(tt.seq)
}

\arguments{
  \item{tt.seq}{a list with objects of class \code{TT.Seq.Gene} or 
 	\code{TT.Seq.Tx}}
}

\value{
Returns Spearman correlation estimates as an vector
}

\author{
Paul K. Korir
}

\examples{
# location of example input files
# to be added...
}

\keyword{correlation}
