\name{numericORlogical-class}
\Rdversion{1.1}
\docType{class}
\alias{numericORlogical-class}

\title{Class \code{"numericORlogical"}}

\description{
A class for objects of numeric OR logical.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "numericORlogical" in the signature.
}

\author{
Paul K. Korir
}

\examples{
## Not run:
showClass("numericORlogical")
## End(Not run)
}

\keyword{classes}
