\name{numericORlogicalORmatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{numericORlogicalORmatrix-class}

\title{Class \code{"numericORlogicalORmatrix"}}
\description{
A class for objects of numeric OR logical OR matrix.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "numericORlogicalORmatrix" in the signature.
}

\author{
Paul K. Korir
}

\examples{
showClass("numericORlogicalORmatrix")
}

\keyword{classes}
