\name{numericORmatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{numericORmatrix-class}

\title{Class \code{"numericORmatrix"}}

\description{
A class for objects of numeric OR matrix.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "numericORmatrix" in the signature.
}

\author{
Paul K. Korir
}

\examples{
showClass("numericORmatrix")
}

\keyword{classes}
